% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFatFreeMassCol.R
\name{calcFatFreeMassCol}
\alias{calcFatFreeMassCol}
\title{calcFatFreeMassCol}
\usage{
calcFatFreeMassCol(df, idVar = "ID", ageVar = "AGE", weightVar = "WT",
  heightVar = "HT", sexVar = "SEX", ageUnit = "years",
  heightConv = 0.01, weightConv = 1, femaleSexVal = 2, maleSexVal = 1,
  childCutoff = 18, missingVal = -99)
}
\arguments{
\item{df}{A data frame with all necessary columns for calculation of fat 
free mass.}

\item{idVar}{Name of the id column in df. Default is "ID".}

\item{ageVar}{Name of the age column in df. Default is "AGE".}

\item{weightVar}{Name of the weight column in df. Default is "WT".}

\item{heightVar}{Name of the height column in df. Default is "HT".}

\item{sexVar}{Name of the sex column in df. Default is "SEX".}

\item{ageUnit}{A single string describing the unit for age in the data. 
"years", "months", or "days" are allowed. Defult is "years"}

\item{heightConv}{Height conversion factor to meters. Height*heightConv 
is height in meters. Default is 0.01, meaning that 
centimeters is expected in data.}

\item{weightConv}{Weight conversion factor to kilograms. Default is 1,
meaning that kilograms is expected in data. There is 
currently no support for stones or pounds.}

\item{femaleSexVal}{The value denoting female sex in the supplied df.
Default is 2.}

\item{maleSexVal}{The value denoting male sex in the supplied df. Default 
is 1.}

\item{childCutoff}{The age, in years, at which a person is considered to be
an adult. Default is 18.}

\item{missingVal}{The value that will be inserted in place of Z-scores that
cannot be determined. Either because age is out of bounds 
of the reference WHO data, or because one of the required
values for age, weight, or height was missing from input 
data. Default is -99 according to Perl speaks NONMEM 
standard.}
}
\description{
Calculates fat free mass for each row in a data frame and 
returns a vector of the same length as the number of columns
in that data frame.
}
